fixing weird thing
